feat: FontRankingChart 컴포넌트 구현 (#41)
